requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
surrealdb>=0.3.0
pydantic>=2.0.0
openai>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
rich>=13.0.0
typer>=0.9.0
uvloop>=0.17.0
asyncio>=3.4.3
httpx>=0.24.0
aiofiles>=23.0.0
loguru>=0.7.0
tenacity>=8.0.0
jinja2>=3.1.0
tiktoken>=0.5.0
